# api/openapi.yaml
openapi: 3.0.3
info:
  title: Velofreundliches Wetzikon — Backend API
  version: 0.1.0
  description: |
    Public + admin API for the velofreundliches-wetzikon backend. Two kinds of consumers:

    1. **Public, unauthenticated** — the Contao CMS site (`velofreundliches-wetzikon-contao`,
       server-side fetch) and the standalone VeloMelder page (`public/additions/velo-melder.html`,
       browser fetch) both read/write the *same* public endpoints below. There is no API key for
       these: they replace what are static files today (`velo-routes.geojson`,
       `velo-meldungen.json`), so anyone who could `curl` the old file can call the new endpoint —
       that was already true and is not a regression.
    2. **Admin, session-authenticated** — the admin backend's own UI (see `../design-prototype/`)
       calls the `/admin/*` paths after a 2FA login. These are never called from the public site
       or the public VeloMelder page.

    Status model for a report ("Meldung" in the German-language UI — see
    `../design-prototype/meldungen.html`/`meldung-detail.html`; naming here is English to match
    the database schema, see `../database/schema.sql`): `pending_email_confirmation` → (email
    link clicked) → `pending_review` → (admin decision) → `published` **or** `declined`. The
    public `GET /reports` list only ever returns `published` entries. See
    `../docs/api-implementation-strategy.md` for the reasoning behind this shape, and
...
    spec is based on.

...
servers:
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Step 1 — email + password
      description: >
        On success, emails a 6-digit code and returns a short-lived challenge token for step 2.
        Does not yet establish a session. There is no separate username: admin accounts are
        identified by email address only (matches how the code in step 2 is delivered).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        '200':
          description: Credentials valid — code sent.
          content:
            application/json:
              schema:
                type: object
                properties:
                  challengeToken: { type: string, description: Opaque token, submit with the code in the next step. }
                  maskedEmail: { type: string, example: "s•••••@velofreundliches-wetzikon.ch" }
        '401':
          description: Invalid email address or password.
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/login/verify:
    post:
      tags: [Auth]
      summary: Step 2 — 6-digit email code
      description: On success, establishes the admin session (sets an httpOnly session cookie).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challengeToken, code]
              properties:
                challengeToken: { type: string }
                code: { type: string, pattern: '^[0-9]{6}$' }
      responses:
        '200':
          description: Session established.
          headers:
            Set-Cookie:
              schema: { type: string, example: "vw_admin_session=…; HttpOnly; Secure; SameSite=Strict" }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminUser' }
        '401':
          description: Code incorrect or challenge token unknown.
        '410':
          description: Code expired — request a new one via /auth/login again.
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request a password-reset code
      description: >
        Always returns 200 regardless of whether the email address exists, to avoid revealing
        which addresses have an account. If it exists, emails a 6-digit code (same mechanism as
        login 2FA).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties: { email: { type: string, format: email } }
      responses:
        '200':
          description: If the account exists, a code was sent.
          content:
            application/json:
              schema:
                type: object
                properties:
                  challengeToken: { type: string }
  /auth/forgot-password/verify:
    post:
      tags: [Auth]
      summary: Confirm the reset code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challengeToken, code]
              properties:
                challengeToken: { type: string }
                code: { type: string, pattern: '^[0-9]{6}$' }
      responses:
        '200':
          description: Code confirmed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  resetToken: { type: string, description: Opaque, single-use, short-lived (e.g. 10 min). }
        '401':
          description: Code incorrect.
        '410':
          description: Code expired.
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Set a new password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resetToken, newPassword]
              properties:
                resetToken: { type: string }
                newPassword: { type: string, format: password, minLength: 12 }
      responses:
        '200':
          description: Password changed. Any existing sessions for this user are invalidated.
        '400':
          description: Password does not meet policy.
        '401':
          description: Reset token invalid, expired, or already used.
  /auth/logout:
    post:
      tags: [Auth]
      summary: End the current admin session
      security: [{ sessionAuth: [] }]
      responses:
        '204': { description: Logged out. }
  /auth/me:
    get:
      tags: [Auth]
      summary: Current admin user
      security: [{ sessionAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminUser' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ---------------------------------------------------------------- ADMIN: REPORTS
...
  /admin/routes:
    get:
      tags: [Admin — Routes]
      summary: Current Velonetz GeoJSON, with each feature's database id
      description: >
        Same data as the public `GET /routes`, but every feature's `properties.id` is included so
        the editor UI can round-trip it — saving with `id` unchanged tells `PUT /admin/routes`
        "this exact line", not "delete this one, draw a new one". `id` is admin-only; the public
        endpoint never includes it.
      security: [{ sessionAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminRouteFeatureCollection' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      tags: [Admin — Routes]
      summary: Replace the full Velonetz GeoJSON
      description: >
        Full-replace at the wire level, matching today's "download from the editor, re-upload by
        hand" workflow — the editor UI keeps its exact current drawing/editing behavior and just
        calls this instead of exporting a file. See data-contract §3.3.

        Server-side, this is a reconciliation, not a blind delete-and-reinsert: features are
        matched to existing rows by `properties.id` (see `GET /admin/routes`), so unchanged and
        edited lines keep their database id across saves — an id-less feature (or one whose id
        doesn't match an existing row) is treated as new. Before writing anything, the full
        pre-change state is snapshotted to a GeoJSON file and recorded as a `RouteBackup` (see
        `GET /admin/routes/backups`), so a bad edit can be rolled back — this replaces the "route
        edit history is out of scope" note that used to be here (strategy doc §7).

        The editor is expected to call `POST /admin/routes/diff` first and show the admin the
        resulting change summary before calling this endpoint — there is no server-side
        distinction between "diffed first" and not, this is a UI-level convention, not enforced
        here.
      security: [{ sessionAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RouteFeatureCollection' }
      responses:
        '200':
          description: Replaced (or a no-op if the proposed state was identical to the current one — no backup is created for a no-op).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteSaveResult' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /admin/routes/diff:
    post:
      tags: [Admin — Routes]
      summary: Preview what saving a proposed GeoJSON would change — no write
      description: >
        Runs the exact same matching/classification `PUT /admin/routes` would, but persists
        nothing and writes no backup. Meant to render a human-readable, bulleted change summary
        (`changeSummary.lines`) for the admin to review before they confirm the save.
      security: [{ sessionAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RouteFeatureCollection' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteChangeSummary' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /admin/routes/backups:
    get:
      tags: [Admin — Routes]
      summary: List pre-change GeoJSON snapshots, newest first
      security: [{ sessionAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/RouteBackup' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /admin/routes/backups/{id}/restore:
    post:
      tags: [Admin — Routes]
      summary: Restore the route network to a previous backup's state
      description: >
        Not a special-cased "undo" — internally this loads the backup's snapshot and runs it
        through the exact same save path as `PUT /admin/routes`, which means restoring itself
        creates a new backup of the state right before the restore. Restoring is therefore always
        reversible too.
      security: [{ sessionAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Restored.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteSaveResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
...
  responses:
    ValidationError:
      description: Request body failed validation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: No valid admin session.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: No resource with that id.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Too many requests from this client.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: vw_admin_session
      description: >
        httpOnly, Secure, SameSite=Strict session cookie set by /auth/login/verify. No API keys
        or bearer tokens for admin endpoints — see docs/api-implementation-strategy.md §2 for why
        session-cookie auth was chosen over a token-based scheme for this use case.
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
...
    RouteFeatureCollection:
      type: object
      properties:
        type: { type: string, enum: [FeatureCollection] }
        features:
          type: array
          items: { $ref: '#/components/schemas/RouteFeature' }
    RouteFeature:
      type: object
      properties:
        type: { type: string, enum: [Feature] }
        geometry:
          type: object
          properties:
            type: { type: string, enum: [LineString] }
            coordinates:
              type: array
              items:
                type: array
                items: { type: number }
                minItems: 2
                maxItems: 2
        properties:
          type: object
          required: [type]
          properties:
            type: { type: string, description: "Key into the ROUTE_TYPES registry, e.g. 'vorschlag-gelbes-band'." }
            direction: { type: string, enum: [one-way, both-ways], description: "Only present for non-band, direction-aware types." }
    AdminRouteFeatureCollection:
      type: object
      properties:
        type: { type: string, enum: [FeatureCollection] }
        features:
          type: array
          items: { $ref: '#/components/schemas/AdminRouteFeature' }
    AdminRouteFeature:
      allOf:
        - $ref: '#/components/schemas/RouteFeature'
        - type: object
          properties:
            properties:
              type: object
              properties:
                id: { type: integer, description: "Database id. Send back unchanged on PUT /admin/routes to update this exact line; omit for a new one." }
    RouteSaveResult:
      allOf:
        - $ref: '#/components/schemas/AdminRouteFeatureCollection'
        - type: object
          properties:
            changeSummary: { $ref: '#/components/schemas/RouteChangeSummary' }
            backupId: { type: integer, nullable: true, description: "Id of the RouteBackup created by this save; null if the save was a no-op (nothing actually changed)." }
    RouteChangeSummary:
      type: object
      properties:
        addedCount: { type: integer }
        removedCount: { type: integer }
        modifiedCount: { type: integer }
        unchangedCount: { type: integer }
        lines:
          type: array
          items: { type: string }
          description: 'Ready-to-render German bullet lines, e.g. "Strecke geändert (Velostreifen, bei 47.3245, 8.7310): Verlauf angepasst (4 → 6 Punkte)".'
        entries:
          type: array
          description: Same information as `lines`, structured (action/label/location/before/after) for a UI that wants to render its own layout instead of the pre-formatted text.
          items: { $ref: '#/components/schemas/RouteChangeEntry' }
    RouteChangeEntry:
      type: object
      properties:
        action: { type: string, enum: [added, removed, modified] }
        id: { type: integer, nullable: true, description: "Null for 'added' (no database id assigned until save)." }
        label: { type: string, description: Route type label the change is grouped under (the new type for added/modified, the old type for removed). }
        location: { type: string, description: "Approximate 'lat, lng' centroid of the line, for a human to recognise which one this is — no reverse geocoding is wired up (see README.md)." }
    RouteBackup:
      type: object
      description: Metadata for one pre-change GeoJSON snapshot. The snapshot file itself isn't exposed over the API; it backs POST /admin/routes/backups/{id}/restore server-side.
      properties:
        id: { type: integer }
        createdAt: { type: string, format: date-time }
        createdByEmail: { type: string, nullable: true }
        addedCount: { type: integer }
        removedCount: { type: integer }
        modifiedCount: { type: integer }
        summary: { type: string, description: The full bulleted change-summary text for the save that created this backup. }
...
    AdminUser:
      type: object
      properties:
        email: { type: string, format: email }
        displayName: { type: string }
        role: { type: string, example: "Redaktion", description: "Not shown in the admin UI's top navigation (see design-prototype/), but still tracked for authorization." }